Split order details into separate view files
